+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Change log
 ##########
 
+1.3.1 (unreleased)
+==================
+
+This release drops support for Python 3.7.
+Python 3.8 or later is now required.
+
+- Require Python 3.8 and drop Python 3.7 support.
+
 1.3.0 (2020-05-19)
 ==================